+++ projects/angular-auth-oidc-client/src/lib/flows/callback-handling/state-validation-callback-handler.service.ts
@@ -21,7 +21,6 @@ export class StateValidationCallbackHandlerService {
   ) {}
 
   // STEP 4 All flows
-
   callbackStateValidation(
     callbackContext: CallbackContext,
     configuration: OpenIdConfiguration,