+++ projects/angular-auth-oidc-client/src/lib/validation/token-validation.service.ts
@@ -4,11 +4,12 @@ import { base64url } from 'rfc4648';
 import { from, Observable, of } from 'rxjs';
 import { map, mergeMap, tap } from 'rxjs/operators';
 import { OpenIdConfiguration } from '../config/openid-configuration';
+import { JwkExtractor } from '../extractors/jwk.extractor';
 import { LoggerService } from '../logging/logger.service';
 import { TokenHelperService } from '../utils/tokenHelper/token-helper.service';
-import { JwtWindowCryptoService } from './jwt-window-crypto.service';
-import { JwkExtractor } from '../extractors/jwk.extractor';
 import { JwkWindowCryptoService } from './jwk-window-crypto.service';
+import { JwtWindowCryptoService } from './jwt-window-crypto.service';
+import { alg2kty, getImportAlg, getVerifyAlg } from './token-validation.helper';
 
 // http://openid.net/specs/openid-connect-implicit-1_0.html
 
@@ -71,29 +72,15 @@ export class TokenValidationService {
 
   // id_token C7: The current time MUST be before the time represented by the exp Claim
   // (possibly allowing for some small leeway to account for clock skew).
-  hasIdTokenExpired(
-    token: string,
-    configuration: OpenIdConfiguration,
-    offsetSeconds?: number,
-    disableIdTokenValidation?: boolean
-  ): boolean {
+  hasIdTokenExpired(token: string, configuration: OpenIdConfiguration, offsetSeconds?: number): boolean {
     const decoded = this.tokenHelperService.getPayloadFromToken(token, false, configuration);
 
-    return !this.validateIdTokenExpNotExpired(decoded, configuration, offsetSeconds, disableIdTokenValidation);
+    return !this.validateIdTokenExpNotExpired(decoded, configuration, offsetSeconds);
   }
 
   // id_token C7: The current time MUST be before the time represented by the exp Claim
   // (possibly allowing for some small leeway to account for clock skew).
-  validateIdTokenExpNotExpired(
-    decodedIdToken: string,
-    configuration: OpenIdConfiguration,
-    offsetSeconds?: number,
-    disableIdTokenValidation?: boolean
-  ): boolean {
-    if (disableIdTokenValidation) {
-      return true;
-    }
-
+  validateIdTokenExpNotExpired(decodedIdToken: string, configuration: OpenIdConfiguration, offsetSeconds?: number): boolean {
     const tokenExpirationDate = this.tokenHelperService.getTokenExpirationDate(decodedIdToken);
 
     offsetSeconds = offsetSeconds || 0;
@@ -338,6 +325,10 @@ export class TokenValidationService {
   // id_token C6: The alg value SHOULD be RS256. Validation of tokens using other signing algorithms is described in the
   // OpenID Connect Core 1.0 [OpenID.Core] specification.
   validateSignatureIdToken(idToken: string, jwtkeys: any, configuration: OpenIdConfiguration): Observable<boolean> {
+    if (!idToken) {
+      return of(true);
+    }
+
     if (!jwtkeys || !jwtkeys.keys) {
       return of(false);
     }
@@ -363,18 +354,16 @@ export class TokenValidationService {
       return of(false);
     }
 
-    const kty = this.alg2kty(alg);
+    const kty = alg2kty(alg);
     const use = 'sig';
 
     try {
-      foundKeys = kid ?
-        this.jwkExtractor.extractJwk(keys, {kid, kty, use}, false) :
-        this.jwkExtractor.extractJwk(keys, {kty, use}, false);
+      foundKeys = kid
+        ? this.jwkExtractor.extractJwk(keys, { kid, kty, use }, false)
+        : this.jwkExtractor.extractJwk(keys, { kty, use }, false);
 
       if (foundKeys.length === 0) {
-        foundKeys = kid ?
-          this.jwkExtractor.extractJwk(keys, {kid, kty}) :
-          this.jwkExtractor.extractJwk(keys, {kty});
+        foundKeys = kid ? this.jwkExtractor.extractJwk(keys, { kid, kty }) : this.jwkExtractor.extractJwk(keys, { kty });
       }
 
       key = foundKeys[0];
@@ -384,7 +373,7 @@ export class TokenValidationService {
       return of(false);
     }
 
-    const algorithm: RsaHashedImportParams | EcKeyImportParams = this.getImportAlg(alg);
+    const algorithm: RsaHashedImportParams | EcKeyImportParams = getImportAlg(alg);
 
     const signingInput = this.tokenHelperService.getSigningInputFromToken(idToken, true, configuration);
     const rawSignature = this.tokenHelperService.getSignatureFromToken(idToken, true, configuration);
@@ -399,7 +388,7 @@ export class TokenValidationService {
       mergeMap((cryptoKey: CryptoKey) => {
         const signature: Uint8Array = base64url.parse(rawSignature, { loose: true });
 
-        const verifyAlgorithm: RsaHashedImportParams | EcdsaParams = this.getVerifyAlg(alg);
+        const verifyAlgorithm: RsaHashedImportParams | EcdsaParams = getVerifyAlg(alg);
 
         return from(this.jwkWindowCryptoService.verifyKey(verifyAlgorithm, cryptoKey, signature, signingInput));
       }),
@@ -409,85 +398,6 @@ export class TokenValidationService {
         }
       })
     );
-  }
-
-  private getImportAlg(alg: string): RsaHashedImportParams | EcKeyImportParams {
-    switch (alg.charAt(0)) {
-      case 'R':
-        if (alg.includes('256')) {
-          return {
-            name: 'RSASSA-PKCS1-v1_5',
-            hash: 'SHA-256',
-          };
-        } else if (alg.includes('384')) {
-          return {
-            name: 'RSASSA-PKCS1-v1_5',
-            hash: 'SHA-384',
-          };
-        } else if (alg.includes('512')) {
-          return {
-            name: 'RSASSA-PKCS1-v1_5',
-            hash: 'SHA-512',
-          };
-        } else {
-          return null;
-        }
-      case 'E':
-        if (alg.includes('256')) {
-          return {
-            name: 'ECDSA',
-            namedCurve: 'P-256',
-          };
-        } else if (alg.includes('384')) {
-          return {
-            name: 'ECDSA',
-            namedCurve: 'P-384',
-          };
-        } else {
-          return null;
-        }
-      default:
-        return null;
-    }
-  }
-
-  private getVerifyAlg(alg: string): RsaHashedImportParams | EcdsaParams {
-    switch (alg.charAt(0)) {
-      case 'R':
-        return {
-          name: 'RSASSA-PKCS1-v1_5',
-          hash: 'SHA-256',
-        };
-      case 'E':
-        if (alg.includes('256')) {
-          return {
-            name: 'ECDSA',
-            hash: 'SHA-256',
-          };
-        } else if (alg.includes('384')) {
-          return {
-            name: 'ECDSA',
-            hash: 'SHA-384',
-          };
-        } else {
-          return null;
-        }
-      default:
-        return null;
-    }
-  }
-
-  private alg2kty(alg: string): string {
-    switch (alg.charAt(0)) {
-      case 'R':
-        return 'RSA';
-
-      case 'E':
-        return 'EC';
-
-      default:
-        throw new Error('Cannot infer kty from alg: ' + alg);
-    }
   }
 
   // Accepts ID Token without 'kid' claim in JOSE header if only one JWK supplied in 'jwks_url'